+++ HISTORY.md
@@ -1,5 +1,8 @@
 # HISTORY
 
+## 27/8/19
+- improve `memoryFindData` function - you can show results that matches multiple checks to easy find relevant memory positions
+
 ## 26/8/19
 - Add memory monitor to UI, reachable via "M" key, "B" and "N" are used to change pages
 - Make sure patched memory is reflected in memory monitor